feat(ui): add MyUpload component for file uploads
